requirements.txt
asgiref==3.6.0
awsebcli==3.20.5
bcrypt==4.0.1
botocore==1.29.81
cement==2.8.2
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
distlib==0.3.6
Django==4.1.7
dnspython==2.2.1
email-validator==1.3.1
filelock==3.8.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.32
pathspec==0.10.1
Pillow==9.4.0
platformdirs==2.6.0
psycopg2-binary==2.9.5
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
SQLAlchemy==2.0.7
sqlparse==0.4.2
termcolor==1.1.0
typing_extensions==4.5.0
tzdata==2022.1
urllib3==1.26.15
virtualenv==20.17.1
wcwidth==0.1.9
Werkzeug==2.2.2
WTForms==3.0.1
